moves renderConnected to react/utils
