+++ app/react/Library/components/specs/TableRow.spec.ts
@@ -3,7 +3,7 @@
  */
 import Immutable from 'immutable';
 
-import { renderConnected } from 'app/Templates/specs/utils/renderConnected';
+import { renderConnected } from 'app/utils/test/renderConnected';
 import { TableRow } from 'app/Library/components/TableRow';
 import { TableCell } from 'app/Library/components/TableCell';
 import { EntitySchema } from 'shared/types/entityType';
