+++ app/react/Templates/components/specs/FormConfigInput.spec.js
@@ -2,7 +2,7 @@ import { Field } from 'react-redux-form';
 import FormConfigInput from '../FormConfigInput';
 import PropertyConfigOptions from '../PropertyConfigOptions';
 
-import { renderConnected } from '../../specs/utils/renderConnected.tsx';
+import { renderConnected } from '../../../utils/test/renderConnected.tsx';
 
 describe('FormConfigInput', () => {
   let props;
